==> motif evolution/scripts/01 - motif table DMI.ipynb <==
### Description:  
This is the first step in the motif conservation analysis (results shown in shown in fig7).

In this step we create 2 tables -> one with dmi motifs and one with non-dmi motifs  
  
**reminder:**  
* "non-dmi" is a sequence that matches a regular expression (regex) pattern of a certain motif that was dropped for 1 or 2 reasons:
1. The motif does not match a domain of the human PPIs of said viral protein
2. The average plddt of said motif is higher than 50 

import pandas as pd
import os
import ast

## motif regex pattern dictionary

elm_motifs_and_regex_clean = # path for elm_motifs_data_regex.tsv #

elm_df = pd.read_csv(elm_motifs_and_regex_clean , delimiter='\t', on_bad_lines='skip')

elm_dict=dict(zip(elm_df.ELMIdentifier, elm_df.Regex))

## HSV1

# Table of all the virus's disordered motifs from the "identify DMIs" script
hsv_disordered_motifs_path = r"hsv_motifs_plddt_disorder_only.csv"

hsv_disorder_motifs_df = pd.read_csv(hsv_disordered_motifs_path )

hsv_disorder_motifs_df.head()

hsv_disorder_motifs_df = hsv_disorder_motifs_df[['v_prot' , 'motif_name']].drop_duplicates()

### add regex

# Using elm's database - creating a column for each motif with its regular expressin
hsv_disorder_motifs_df['motif_regex'] = hsv_disorder_motifs_df['motif_name'].map(elm_dict)

hsv_disorder_motifs_df.head()

### is said motif in dmi data or not?

motifs in dmi table are already only those who appear in disordered regions

hsv_dmi_path = r"hsv_dmi_table.csv"

hsv_dmis_df = pd.read_csv(hsv_dmi_path)

hsv_dmis_df.head()

# Create dictionary of the dmi table from the "identify DMIs"

hsv_dmis_dict = {}

for prot, sub_df in hsv_dmis_df.groupby('v_prot'):
    # Convert stringified lists to actual lists
    motif_series = sub_df['motif'].apply(ast.literal_eval)

    # Flatten the list and remove duplicates
    flat_unique_motifs = list(set(motif for sublist in motif_series for motif in sublist))

    # Save to dictionary
    hsv_dmis_dict[prot] = flat_unique_motifs

# Function recieve a row of motifs table- looks for it in the dmi dictionary
# Defines motif as DMI or non DMI

def is_dmi(row):
    prot = row['v_prot']
    motif = row['motif_name']
    if prot in hsv_dmis_dict.keys():
        if motif in hsv_dmis_dict.get(prot):
            return "dmi"
        else:
            return "non dmi"
    else:
        return "non dmi"

# Add column "is_dmi" to define motifs as dmi/non dmi
hsv_disorder_motifs_df['is_dmi'] = hsv_disorder_motifs_df.apply(is_dmi, axis=1)

hsv_disorder_motifs_df.head()

### split to 2 tables - dmi motifs and non-dmi motifs

hsv_motifs_df_dmi = hsv_disorder_motifs_df[hsv_disorder_motifs_df['is_dmi'] == 'dmi']
hsv_motifs_df_non_dmi = hsv_disorder_motifs_df[hsv_disorder_motifs_df['is_dmi'] == 'non dmi']

hsv_motifs_df_dmi.to_csv(r'hsv_dmi_motifs.csv', index= False)

hsv_motifs_df_non_dmi.to_csv(r'hsv_non_dmi_motifs.csv', index= False)
